Rename test files to be more descriptive
